refactor: modal창 코드 분리 (js) #51
